requirements.txt
anthropic
requests
jinja2
pyyaml
python-dotenv
